+++ abr_algo_standalone/myant/data_processing.h
@@ -6,7 +6,6 @@
 #include <stdint.h>
 #include <string.h>
 
-
-double digital_filter(double input, double *in, double *out, const double *a, const double *b, uint8_t a_len, uint8_t b_len, uint8_t filter_order, bool reset, double initial_sample);
+float digital_filter(float input, float *in, float *out, const float *a, const float *b, uint8_t a_len, uint8_t b_len, uint8_t filter_order, bool reset, float initial_sample);
 
 #endif /* DATA_PROCESSING_H_ */
